Remove unused imports, added documentation
