--- Domain/sdLitica.Analytics/Abstractions/IAnalyticsOperation.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace sdLitica.Analytics.Abstractions
-{
-    public interface IAnalyticsOperation
-    {
-        string OpName { get; set; }
-        string TsId { get; set; }
-        List<OperationParameter> Parameters { get; set; } // IOperationParameter
-        
-
-    }
-}
