--- Domain/sdLitica.Analytics/Abstractions/IOperationParameter.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace sdLitica.Analytics.Abstractions
-{
-    public interface IOperationParameter
-    {
-        string Name { get; }
-        string Value { get; }
-    }
-}
