--- Domain/sdLitica.Analytics/Abstractions/OperationParameter.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace sdLitica.Analytics.Abstractions
-{
-    public class OperationParameter: IOperationParameter
-    {
-        public OperationParameter()
-        {
-
-        }
-        public string Name { get; set; }
-        public string Value { get; set; }
-    }
-}
