+++ Domain/sdLitica.Analytics/OperationParameter.cs
@@ -1,0 +1,17 @@
+﻿namespace sdLitica.Analytics
+{
+
+    //Not used at this moment.Probably will be replaced with simple dictionary field. 
+
+    /// <summary>
+    /// Single parameter of AnalyticsOperation.
+    /// </summary>
+    public class OperationParameter
+    {        public OperationParameter()
+        {
+
+        }
+        public string Name { get; set; }
+        public string Value { get; set; }
+    }
+}
